+++ src/it/cnr/isti/vir/util/L1.java
@@ -54,18 +54,18 @@ public class L1 {
 	}
 	
 	
-	public static float get(float[] v1, float[] v2) {
-		float dist = 0;		
+	public static double get(float[] v1, float[] v2) {
+		double dist = 0;		
 	    for (int i = 0; i < v1.length; i++) {
-	    	dist += Math.abs(v1[i] - v2[i]);
+	    	dist += Math.abs((double) v1[i] - v2[i]);
 	    }		
 		return dist;
 	}
 	
-	public static float get(float[] v1, float[] v2, float maxDist) {
+	public static double get(float[] v1, float[] v2, float maxDist) {
 		float dist = 0;
 	    for (int i = 0; i < v1.length; i++) {
-	    	dist += Math.abs(v1[i] - v2[i]);
+	    	dist += Math.abs((double) v1[i] - v2[i]);
 	    	if ( dist > maxDist ) return -dist;
 	    }		
 		return dist;
